+++ app/components/developers/query_component.rb
@@ -60,6 +60,19 @@ module Developers
       query.include_not_interested
     end
 
+    def collapse_location?
+      query.countries.empty?
+    end
+
+    def collapse_all_locations?
+      @collapse_all_locations ||=
+        (Location.not_top_countries & query.countries).none?
+    end
+
+    def collapse_timezone?
+      query.utc_offsets.empty?
+    end
+
     private
 
     def utc_offsets