+++ app/components/users/paywalled_component.rb
@@ -1,15 +1,20 @@
 module Users
   class PaywalledComponent < ApplicationComponent
-    def initialize(user:, paywalled:, size: nil, title: nil, description: nil)
+    renders_one :custom_paywall
+
+    private attr_reader :user, :resource, :public_key
+
+    def initialize(user, resource, size: nil, title: nil, description: nil, public_key: nil)
       @user = user
-      @paywalled = paywalled
+      @resource = resource
       @size = size
       @title = title
       @description = description
+      @public_key = public_key
     end
 
     def render_content?
-      customer? || owner?
+      Users::Permission.new(user, resource, public_key:).authorized?
     end
 
     def small?
@@ -26,16 +31,6 @@ module Users
 
     def description
       @description || t(".description")
-    end
-
-    private
-
-    def customer?
-      Businesses::Permission.new(@user&.subscriptions).active_subscription?
-    end
-
-    def owner?
-      @paywalled&.user == @user && @user.present?
     end
   end
 end