+++ app/controllers/cold_messages_controller.rb
@@ -23,7 +23,7 @@ class ColdMessagesController < ApplicationController
   private
 
   def cold_message(message)
-    ColdMessage.new(message:, show_hiring_fee_terms: permission.pays_hiring_fee?)
+    ColdMessage.new(message:, show_hiring_fee_terms: permissions.pays_hiring_fee?)
   end
 
   def require_business!
@@ -38,7 +38,7 @@ class ColdMessagesController < ApplicationController
   end
 
   def require_active_subscription!
-    unless permission.active_subscription?
+    unless permissions.active_subscription?
       store_location!
       redirect_to pricing_path
     end
@@ -51,8 +51,8 @@ class ColdMessagesController < ApplicationController
     end
   end
 
-  def permission
-    @permission = Businesses::Permission.new(current_user.subscriptions)
+  def permissions
+    @permissions = current_user.permissions
   end
 
   def conversation