+++ db/migrate/20220919170633_add_public_profile_key_to_developers.rb
@@ -1,0 +1,6 @@
+class AddPublicProfileKeyToDevelopers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :developers, :public_profile_key, :string
+    add_index :developers, :public_profile_key, unique: true
+  end
+end