+++ db/migrate/20221007145250_remove_manual_flag_to_open_startup_revenues.rb
@@ -1,0 +1,5 @@
+class RemoveManualFlagToOpenStartupRevenues < ActiveRecord::Migration[7.0]
+  def change
+    remove_column :open_startup_revenue, :manual, :boolean, null: false, default: false
+  end
+end