+++ db/schema.rb
@@ -10,8 +10,9 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_09_19_191456) do
+ActiveRecord::Schema[7.0].define(version: 2022_10_07_145250) do
   # These are extensions that must be enabled in order to support this database
+  enable_extension "pg_stat_statements"
   enable_extension "plpgsql"
 
   create_table "action_mailbox_inbound_emails", force: :cascade do |t|
@@ -111,6 +112,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_19_191456) do
     t.boolean "profile_reminder_notifications", default: true
     t.string "stack_overflow"
     t.boolean "source_contributor", default: false, null: false
+    t.string "public_profile_key"
+    t.index ["public_profile_key"], name: "index_developers_on_public_profile_key", unique: true
     t.index ["textsearchable_index_col"], name: "textsearchable_index", using: :gin
     t.index ["user_id"], name: "index_developers_on_user_id"
   end
@@ -243,7 +246,6 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_19_191456) do
     t.decimal "amount", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.boolean "manual", default: false, null: false
   end
 
   create_table "open_startup_stripe_transactions", force: :cascade do |t|
@@ -292,7 +294,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_19_191456) do
     t.string "processor_id"
     t.boolean "default"
     t.jsonb "data"
-    t.datetime "deleted_at"
+    t.datetime "deleted_at", precision: nil
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["owner_type", "owner_id", "deleted_at", "default"], name: "pay_customer_owner_index"
@@ -329,8 +331,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_19_191456) do
     t.string "processor_plan", null: false
     t.integer "quantity", default: 1, null: false
     t.string "status", null: false
-    t.datetime "trial_ends_at"
-    t.datetime "ends_at"
+    t.datetime "trial_ends_at", precision: nil
+    t.datetime "ends_at", precision: nil
     t.decimal "application_fee_percent", precision: 8, scale: 2
     t.jsonb "metadata"
     t.jsonb "data"