+++ CHANGELOG.md
@@ -2,8 +2,19 @@
 
 ## 2022
 
-## September
+### October
 
+* October 21 - Option to ignore pages from being tracked in Fathom #706
+* October 20 - Upgrade Tailwind CSS to v3.2 #700
+* October 19 - Upgrade to Turbo 7.2.2 #683 @sarvaiyanidhi
+* October 12 - Remove manual payments and account for payments made via ACH #681
+* October 12 - Update to rails 7.0.4 #682 @kaushikhande
+* October 7 - Bring back 1x and 2x avatar variants and remove lazy image loading #680 @jkostolansky
+
+### September
+
+* September 22  - Display developer private information with publicly shared URL #646 @sarvaiyanidhi
+* September 21 - Only show page 1 of search results to non-subscribers #661 @mansakondo
 * September 18 - Add badges to developer profile pages #657
 * September 18 - Obfuscate developer profiles URLs #656
 * September 16 - Show "Recently Active" badge for developers active in last 7 days #640 @sarvaiyanidhi
@@ -13,10 +24,10 @@
 * September 13 - Impersonate users directly from `/admin/users` #627 @KarlHeitmann
 * September 13 - Move developer bio to own card in form and add more instructions #635 @rayhanw
 * September 8 - Send admins a notification when a `Hired::Form` is created #614 @sarvaiyanidhi
-* September 8 – Update license year and note non-code usage #630
-* September 8 – Remove part-time hiring plan from pricing page #628
+* September 8 – Update license year and note non-code usage #630
+* September 8 – Remove part-time hiring plan from pricing page #628
 
-## August
+### August
 
 * August 27 - Add a URL to Admin::SubscriptionChangeNotification #612
 * August 27 - Fixes an issue where Safari would never load retina images #619